refactor signature
