+++ src/de.rs
@@ -1,0 +1,5 @@
+use errors::Result;
+
+pub trait Deserialize: Sized {
+    fn from_slice(&[u8]) -> Result<Self>;
+}
